+++ src/http/handler/discovery_types.go
@@ -1,8 +1,9 @@
 package handler
 
 type DiscoveryRequest struct {
-	CheckURL string `json:"check_url,omitempty"`
-	SkipDNS  bool   `json:"skip_dns,omitempty"`
+	CheckURL     string   `json:"check_url,omitempty"`
+	SkipDNS      bool     `json:"skip_dns,omitempty"`
+	PayloadFiles []string `json:"payload_files,omitempty"`
 }
 
 type DiscoveryResponse struct {